requirement.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-google-genai
langchain-huggingface
langchain-chroma
langchain-text-splitters
pymupdf
pymupdf4llm
chromadb
sentence-transformers
rank_bm25